moved reasoner abstract classes to new module
